+++ pocketlab/commands/connect.py
@@ -16,7 +16,7 @@ _connect_details = {
 
 from pocketlab.init import fields_model
 
-def connect(platform_name, service_option, environment_type='', resource_tag='', region_name='', verbose=True):
+def connect(platform_name, service_option, environ_type='', resource_tag='', region_name='', verbose=True):
 
     title = 'connect'
 
@@ -28,7 +28,7 @@ def connect(platform_name, service_option, environment_type='', resource_tag='',
         'service_option': service_option,
         'verbose': verbose,
         'platform_name': platform_name,
-        'environment_type': environment_type,
+        'environ_type': environ_type,
         'resource_tag': resource_tag
     }
     for key, value in input_fields.items():
@@ -83,7 +83,7 @@ def connect(platform_name, service_option, environment_type='', resource_tag='',
             service_insert=service_insert, 
             service_root=service_root, 
             region_name=region_name, 
-            environment_type=environment_type, 
+            environment_type=environ_type, 
             resource_tag=resource_tag, 
             verbose=verbose
         )