+++ pocketlab/commands/get.py
@@ -16,7 +16,7 @@ _get_details = {
 
 from pocketlab.init import fields_model
 
-def get(file_path, platform_name, service_option, environment_type='', resource_tag='', region_name='', verbose=True, overwrite=False):
+def get(file_path, platform_name, service_option, environ_type='', resource_tag='', region_name='', verbose=True, overwrite=False):
     
     title = 'get'
 
@@ -29,7 +29,7 @@ def get(file_path, platform_name, service_option, environment_type='', resource_
         'service_option': service_option,
         'verbose': verbose,
         'platform_name': platform_name,
-        'environment_type': environment_type,
+        'environ_type': environ_type,
         'resource_tag': resource_tag
     }
     for key, value in input_fields.items():
@@ -80,7 +80,7 @@ def get(file_path, platform_name, service_option, environment_type='', resource_
             service_insert=service_insert, 
             service_root=service_root, 
             region_name=region_name, 
-            environment_type=environment_type, 
+            environment_type=environ_type, 
             resource_tag=resource_tag, 
             verbose=verbose
         )